working database persistence implementation
this is a fully working implementation of database persistence rather
than the cludge of hashmaps and the config file that I had been using.
there is likely more to be done in cleaning up this implementation, like
relating the two tables I am using, cleaning up the access methods,
etc., but this is a big improvement from before. note; in this commit,
the commands for reporting player deaths and stuff are still
unimplemented. also keeping track of days alive.